Add files via upload
Images and a movie for exploring the covid data.  The movie shows the covid transmission levels for each week on a map.  It's helpful to see the ups and downs and the similar/differences across the state.  The plots I'm trying to explore the relationship between covid transmission level and the two covariates.  So a lot of the plots have "before the estimated change point" and "after the estimated change point" -- I'm trying to see if there is a different relationship between the probabilities of different transmission levels before and after the change point.  Lots of info here.  I'd love for you guys to see what things you take away from the plots